Island: Extract jwt_required decorator into separate file
