+++ monkey/monkey_island/cc/resources/agent_controls/stop_all_agents.py
@@ -3,7 +3,7 @@ import json
 from flask import make_response, request
 
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.resources.utils.semaphores import agent_killing_mutex
 from monkey_island.cc.services.infection_lifecycle import set_stop_all, should_agent_die
