+++ monkey/monkey_island/cc/resources/auth/auth.py
@@ -6,6 +6,7 @@ from flask import make_response, request
 from common.utils.exceptions import IncorrectCredentialsError
 from monkey_island.cc.resources.AbstractResource import AbstractResource
 from monkey_island.cc.resources.auth.credential_utils import get_username_password_from_request
+from monkey_island.cc.resources.request_authentication import create_access_token
 from monkey_island.cc.services import AuthenticationService
 
 logger = logging.getLogger(__name__)
@@ -39,30 +40,8 @@ class Authenticate(AbstractResource):
 
         try:
             AuthenticationService.authenticate(username, password)
-            access_token = _create_access_token(username)
+            access_token = create_access_token(username)
         except IncorrectCredentialsError:
             return make_response({"error": "Invalid credentials"}, 401)
 
         return make_response({"access_token": access_token, "error": ""}, 200)
-
-
-def _create_access_token(username):
-    access_token = flask_jwt_extended.create_access_token(identity=username)
-    logger.debug(f"Created access token for user {username} that begins with {access_token[:4]}")
-
-    return access_token
-
-
-# See https://flask-jwt-extended.readthedocs.io/en/stable/custom_decorators/
-def jwt_required(fn):
-    @wraps(fn)
-    def wrapper(*args, **kwargs):
-        try:
-            flask_jwt_extended.verify_jwt_in_request()
-            return fn(*args, **kwargs)
-        # Catch authentication related errors in the verification or inside the called function.
-        # All other exceptions propagate
-        except (JWTExtendedException, PyJWTError) as e:
-            return make_response({"error": f"Authentication error: {str(e)}"}, 401)
-
-    return wrapper
