+++ monkey/monkey_island/cc/resources/blackbox/clear_caches.py
@@ -3,7 +3,7 @@ import logging
 import flask_restful
 
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.attack.attack_report import AttackReportService
 from monkey_island.cc.services.reporting.report import ReportService
