+++ monkey/monkey_island/cc/resources/blackbox/log_blackbox_endpoint.py
@@ -3,7 +3,7 @@ from flask import request
 
 from monkey_island.cc.database import database, mongo
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 
 
 class LogBlackboxEndpoint(AbstractResource):
