+++ monkey/monkey_island/cc/resources/configuration_import.py
@@ -7,7 +7,7 @@ from flask import request
 
 from common.utils.exceptions import InvalidConfigurationError
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.server_utils.encryption import (
     InvalidCiphertextError,
     InvalidCredentialsError,
