+++ monkey/monkey_island/cc/resources/exploitations/manual_exploitation.py
@@ -1,5 +1,5 @@
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.reporting.exploitations.manual_exploitation import (
     get_manual_exploitations,
 )
