+++ monkey/monkey_island/cc/resources/island_logs.py
@@ -1,7 +1,7 @@
 import logging
 
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.island_logs import IslandLogService
 
 logger = logging.getLogger(__name__)
