+++ monkey/monkey_island/cc/resources/local_run.py
@@ -4,7 +4,7 @@ from flask import jsonify, make_response, request
 
 from monkey_island.cc.models import Monkey
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.node import NodeService
 from monkey_island.cc.services.run_local_monkey import LocalMonkeyRunService
