+++ monkey/monkey_island/cc/resources/node.py
@@ -1,7 +1,7 @@
 from flask import request
 
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.node import NodeService
