+++ monkey/monkey_island/cc/resources/node_states.py
@@ -1,5 +1,5 @@
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.utils.node_states import NodeStates as NodeStateList
