+++ monkey/monkey_island/cc/resources/ransomware_report.py
@@ -1,7 +1,7 @@
 from flask import jsonify
 
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.ransomware import ransomware_report
