+++ monkey/monkey_island/cc/resources/remote_run.py
@@ -5,7 +5,7 @@ from botocore.exceptions import ClientError, NoCredentialsError
 from flask import jsonify, make_response, request
 
 from monkey_island.cc.resources.AbstractResource import AbstractResource
-from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services import AWSService
 from monkey_island.cc.services.aws import AWSCommandResults
